Split finalizer related code
